+++ packages/components/src/components/select/select.stories.ts
@@ -272,8 +272,7 @@ export const Slots = {
   `tag`,
   `tag__base`,
   `tag__content`,
-  `tag__remove-button`,
-  `tag__remove-button__base`,
+  `tag__removable-indicator`,
   `clear-button`, and
   `expand-icon` part selectors to customize the select component.
  */
@@ -290,8 +289,7 @@ const partsArr = [
   'tag',
   'tag__base',
   'tag__content',
-  'tag__remove-button',
-  'tag__remove-button__base',
+  'tag__removable-indicator',
   'clear-button',
   'expand-icon'
 ];
@@ -312,8 +310,7 @@ export const Parts = {
         'tag',
         'tag__base',
         'tag__content',
-        'tag__remove-button',
-        'tag__remove-button__base',
+        'tag__removable-indicator',
         'clear-button',
         'expand-icon'
       ]
@@ -324,13 +321,14 @@ export const Parts = {
 
     return generateTemplate({
       axis: {
+        x: { type: 'attribute', name: 'useTags' },
         y: {
           type: 'template',
           name: 'sd-select::part(...){outline: solid 2px red}',
           values: partsArr.map(part => {
             return {
               title: part,
-              value: `<style>#part-${part} sd-select::part(${part}){outline: solid 2px red}</style><div id="part-${part}">%TEMPLATE%</div>`
+              value: `<style>#part-${part} sd-select::part(${part}){outline: solid 2px red; outline-offset: 2px} ::part(popup__content){overflow-y: visible} ::part(tag__removable-indicator){display: block} ::part(tag__content){display: block}</style><div id="part-${part}">%TEMPLATE%</div>`
             };
           })
         }
@@ -493,6 +491,72 @@ export const SampleForm = {
         }
 
         document.querySelector('#testForm').addEventListener('submit', handleSubmit);
+      </script>
+    `;
+  }
+};
+
+/**
+ * 1. You can use the `setCustomValidity` method to set a custom validation message. This will override any native validation messages.
+ * 2. Set an empty string to clear the custom validity and make the input valid.
+ * 3. To show the validation message, call the `reportValidity` method. Originally this would show a native validation bubble, but we show the error messages inline.
+ */
+
+export const setCustomValidity = {
+  parameters: {
+    chromatic: { disableSnapshot: true }
+  },
+  render: () => {
+    return html`
+      <!-- block submit and show alert instead -->
+      <form id="validationForm" class="flex flex-col gap-2">
+        <sd-select id="custom-input">
+          <sd-option value="option-1">Option 1</sd-option>
+          <sd-option value="option-2">Option 2</sd-option>
+          <sd-option value="option-3">Option 3</sd-option>
+        </sd-select>
+        <div>
+          <sd-button type="submit">Submit</sd-button>
+          <sd-button id="error-button" variant="secondary">Set custom error</sd-button>
+          <sd-button id="success-button" variant="secondary">Set success</sd-button>
+          <sd-button type="reset" variant="secondary">Reset</sd-button>
+        </div>
+      </form>
+      <script type="module">
+        // Wait for custom elements to be defined
+        await Promise.all([
+          customElements.whenDefined('sd-select'),
+          customElements.whenDefined('sd-button'),
+          customElements.whenDefined('sd-option')
+        ]).then(() => {
+          const form = document.getElementById('validationForm');
+          const input = document.getElementById('custom-input');
+          const setErrorButton = document.getElementById('error-button');
+          const setSuccessButton = document.getElementById('success-button');
+
+          // Initial error
+          const errorMessage = \`This is an initial custom error (\${new Date().toLocaleTimeString()})\`;
+          input.setCustomValidity(errorMessage);
+          input.reportValidity();
+
+          // Show error message
+          setErrorButton.addEventListener('click', () => {
+            const errorMessage = \`This is a new custom error (\${new Date().toLocaleTimeString()})\`;
+            input.setCustomValidity(errorMessage);
+            input.reportValidity();
+          });
+
+          // Show success message
+          setSuccessButton.addEventListener('click', () => {
+            input.setCustomValidity(''); // Clear custom validity
+            input.reportValidity();
+          });
+
+          form.addEventListener('submit', event => {
+            event.preventDefault();
+            alert('All fields are valid!');
+          });
+        });
       </script>
     `;
   }