+++ packages/components/src/components/select/select.ts
@@ -746,6 +746,7 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
 
   /** Checks for validity and shows the browser's validation message if the control is invalid. */
   reportValidity() {
+    this.formControlController.fakeUserInteraction();
     return this.valueInput.reportValidity();
   }
 
@@ -876,6 +877,9 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
             sync="width"
             auto-size="vertical"
             auto-size-padding="10"
+            exportparts="
+              popup:popup__content,
+            "
           >
             <div
               part="combobox"
@@ -961,12 +965,18 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
                 : ''}
               ${this.showInvalidStyle
                 ? html`
-                    <sd-icon class=${cx('text-error', iconMarginLeft, iconSize)} library="system" name="risk"></sd-icon>
+                    <sd-icon
+                      part="invalid-icon"
+                      class=${cx('text-error', iconMarginLeft, iconSize)}
+                      library="system"
+                      name="risk"
+                    ></sd-icon>
                   `
                 : ''}
               ${this.showValidStyle
                 ? html`
                     <sd-icon
+                      part="valid-icon"
                       class=${cx('text-success', iconMarginLeft, iconSize)}
                       library="system"
                       name="confirm"
@@ -1034,7 +1044,7 @@ export default class SdSelect extends SolidElement implements SolidFormControl {
       }
 
       sd-popup::part(popup) {
-        @apply overflow-y-scroll;
+        @apply overflow-y-scroll z-dropdown;
       }
 
       sd-tag::part(base) {