+++ packages/components/CHANGELOG.md
@@ -1,3 +1,58 @@
+# [@solid-design-system/components-v1.38.3](https://github.com/solid-design-system/solid/compare/components/1.38.2...components/1.38.3) (2024-01-19)
+
+
+### Bug Fixes
+
+* 🤔 parts not marked (red border) in storybook in sd-select ([#690](https://github.com/solid-design-system/solid/issues/690)) ([d481213](https://github.com/solid-design-system/solid/commit/d481213f976c682c3bad844e5b1ed95ec5e6f0cf))
+
+### 📈 Stats
+* Bundle size (uncompressed): 254 KB (unchanged)
+* Bundle size (gzipped): 62 KB (unchanged)
+
+# [@solid-design-system/components-v1.38.2](https://github.com/solid-design-system/solid/compare/components/1.38.1...components/1.38.2) (2024-01-18)
+
+
+### Bug Fixes
+
+* 🤔 z-index in sd-select ([#685](https://github.com/solid-design-system/solid/issues/685)) ([8b98607](https://github.com/solid-design-system/solid/commit/8b986078f10a6512155632946dec8e5f039669b5)), closes [#534](https://github.com/solid-design-system/solid/issues/534)
+
+### 📈 Stats
+* Bundle size (uncompressed): 254 KB (unchanged)
+* Bundle size (gzipped): 62 KB (unchanged)
+
+# [@solid-design-system/components-v1.38.1](https://github.com/solid-design-system/solid/compare/components/1.38.0...components/1.38.1) (2024-01-18)
+
+
+### Bug Fixes
+
+* add missing navigation-item parts ([#693](https://github.com/solid-design-system/solid/issues/693)) ([34eb020](https://github.com/solid-design-system/solid/commit/34eb020f505101831df1118c16379ac707f4da36))
+
+### 📈 Stats
+* Bundle size (uncompressed): 254 KB (unchanged)
+* Bundle size (gzipped): 62 KB (unchanged)
+
+# [@solid-design-system/components-v1.38.0](https://github.com/solid-design-system/solid/compare/components/1.37.3...components/1.38.0) (2024-01-15)
+
+
+### Features
+
+* ✨ add interactive teaser sample ([#670](https://github.com/solid-design-system/solid/issues/670)) ([d292cfe](https://github.com/solid-design-system/solid/commit/d292cfef1f63aa98e55b77ae6eabd455e42a7a9c))
+
+### 📈 Stats
+* Bundle size (uncompressed): 254 KB (unchanged)
+* Bundle size (gzipped): 62 KB (unchanged)
+
+# [@solid-design-system/components-v1.37.3](https://github.com/solid-design-system/solid/compare/components/1.37.2...components/1.37.3) (2024-01-10)
+
+
+### Bug Fixes
+
+* 🐛 make setCustomValidity() + reportValidity() work with inputs' inline errors ([#667](https://github.com/solid-design-system/solid/issues/667)) ([455f7f6](https://github.com/solid-design-system/solid/commit/455f7f685c7f0bf7f016e2575f18b77a71b4e7ce))
+
+### 📈 Stats
+* Bundle size (uncompressed): 254 KB (+1 KB / +0%)
+* Bundle size (gzipped): 62 KB (unchanged)
+
 # [@solid-design-system/components-v1.37.2](https://github.com/solid-design-system/solid/compare/components/1.37.1...components/1.37.2) (2024-01-08)